Revert "changing asset exchange interop cc as package"
This reverts commit 606905eedecefdf0f3f48d42300784883a0c8c11.
