+++ core/network/fabric-interop-cc/contracts/interop/Makefile
@@ -1,6 +1,2 @@
-assetexchange:
-	cp -r ../../libs/assetexchange .
-test-manage-assets: assetexchange
+test-manage-assets:
 	go test manage_assets.go manage_assets_test.go main.go setup_test.go -v
-clean:
-	rm -rf assetexchange
