--- core/network/fabric-interop-cc/libs/assetexchange/go.mod
@@ -1,11 +1,0 @@
-module github.com/hyperledger-labs/weaver-dlt-interoperability/core/network/fabric-interop-cc/libs/assetexchange
-
-go 1.16
-
-require (
-	github.com/golang/protobuf v1.5.2
-	github.com/hyperledger-labs/weaver-dlt-interoperability/common/protos-go v1.2.1
-	github.com/hyperledger/fabric-contract-api-go v1.1.1
-	github.com/hyperledger/fabric-protos-go v0.0.0-20210720123151-f0dc3e2a0871
-	github.com/sirupsen/logrus v1.8.1
-)
